rendered active date off by one selection after date change
When the date is changed in the datepicker, the displayed active date is the previous selection in most cycles, it looks like when the date is a change it is not applying the change to the internal `activeDate` that is used for checking for the active date to display as selected. Solution was to ensure that it gets set, and set it in a `$timeout` as it gets missed in most cycles when not in a `$timeout`.